refactor: Abstract application config and starting
Create the `AppConfig` type to abstract the application configuration.

Rename the package `grpc` to `grpc_api` in api/ to avoid conflicts with
the package `grpc` from "google.golang.org/grpc".

Create the `GetUnaryMiddlewares()` function in api/grpc_api/middleware
to abstract the configuration of gRPC middlewares/interceptors.

Create the `StartApp()` function in api/grpc_api to abstract the
application starting.